Fixes #3: hangs when run after compiling with arm-none-eabi-gcc v7.3.1
Debugging showed that const data (in the .rodata section) was corrupted. Data that should be found at a particular address was actually four bytes prior. The DMA init in `DMA_CalcBaseAndBitshift()` would freeze because it uses a static array `flagBitshiftOffset`, which had corrupted data.

The problem was that the linker script allowed the LMA and VMA counters to get out of sync. Adding the ALIGN_WITH_INPUT flag fixed this. See: https://sourceware.org/binutils/docs-2.35/ld/Forced-Output-Alignment.html